requirements.txt
astroid==2.5.6
certifi==2020.12.5
chardet==4.0.0
crc32c==2.2
idna==2.10
isort==5.8.0
kafka-python==2.0.2
lazy-object-proxy==1.6.0
lz4==3.1.3
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.8.2
pylint-venv==2.1.1
pyparsing==2.4.7
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
